Create extractValidSubSeqs.pl
This is a script to correct noisy data from the RoC-ITS sequencing process. It takes clustering/tree data from existing sets of sequences and identifies outliers that suggest that there are multiple inserts in a RoC-ITS read. If the data is consistent with multiple inserts, the data is broken up into independent read sets.