Create analyzeTreeData.pl
Takes the data from analyzeTrees.R and identifies the best potential iteration (if one exists at all) between the potential sub-trees. If the reads consistently bounce between sub-tree A and B (so A,B,A,B,A,B,A,B etc) then this determines whether two (or more) inserts are associated with a given RoC-ITS nanopore read. The input is used by the extractValidSubSeqs.pl script. The output is a tab-delimited file:
1) read id
2) optimal number of inserts
3) number of successes on binomial given number of inserts
4) number of sub-sequences tested